requirements.txt
streamlit==1.32.0
pandas==2.2.0
Pillow==10.2.0
python-dotenv==1.0.0
flask==2.2.3
flask-sqlalchemy==3.0.3
werkzeug==2.2.3
markupsafe==2.1.2
gunicorn==20.1.0